feat: [SCRUM-119,120] update file ultils
