+++ src/shared/lib/constants/colors.js
